+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/structure/Structure.java
@@ -97,6 +97,7 @@ public class Structure {
      * @param labelAsymId the chain-ID of the residue
      * @param labelSeqId the sequence position of the residue
      * @return the index of the residue
+     * @throws NoSuchElementException if the chain/residue cannot be found
      */
     public int getResidueIndex(String labelAsymId, int labelSeqId) {
         if (!chainOffsets.containsKey(labelAsymId)) {
@@ -105,7 +106,11 @@ public class Structure {
 
         int[] chainOffset = chainOffsets.get(labelAsymId);
         // on the sub-array binary search works
-        return Arrays.binarySearch(this.labelSeqId, chainOffset[0], chainOffset[1] + 1, labelSeqId);
+        int index = Arrays.binarySearch(this.labelSeqId, chainOffset[0], chainOffset[1] + 1, labelSeqId);
+        if (index < 0) {
+            throw new NoSuchElementException("Didn't find residue with label_seq_id " + labelSeqId + " in chain " + labelAsymId);
+        }
+        return index;
     }
 
     public LabelSelection getLabelSelection(int residueIndex) {