+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Strucmotif-search Changelog
 =============
 
+strucmotif-search 0.12.2
+-------------
+### Bug fix
+- throw `IllegalQueryDefinitionException` if requested chain/residue isn't found, part 2
+
+strucmotif-search 0.12.1
+-------------
+### Bug fix
+- throw `IllegalQueryDefinitionException` if requested chain/residue isn't found
+
 strucmotif-search 0.12.0
 -------------
 ### Breaking changes